Crate test_mmap

Constants

MAPLEN

Functions

check_buf

init_buf

main

mmap_and_init_buf

page_size

test_mmap_anon

test_mmap_file

test_mmap_file_high32

test_mmap_file_high64

test_mmap_file_low

test_mmap_prot_none_mprotect

test_mremap_clobber

test_munmap_back

Exercise the MemoryManager logic for unmapping the back of a mapped region, validating that the still-mapped part of the region is still accessible. Regression test for #1188.

test_munmap_front

Exercise the MemoryManager logic for unmapping the front of a mapped region, validating that the still-mapped part of the region is still accessible. Regression test for #1188.

test_munmap_middle

Exercise the MemoryManager logic for unmapping the middle of a mapped region, validating that the still-mapped parts of the region is still accessible. Regression test for #1188.

validate_shadow_access