Crate test_mmap

Constants

MAPLEN ğŸ”’

Functions

check_buf ğŸ”’

init_buf ğŸ”’

main ğŸ”’

mmap_and_init_buf ğŸ”’

page_size ğŸ”’

test_mmap_anon ğŸ”’

test_mmap_file ğŸ”’

test_mmap_file_high32 ğŸ”’

test_mmap_file_high64 ğŸ”’

test_mmap_file_low ğŸ”’

test_mmap_prot_none_mprotect ğŸ”’

test_mremap_clobber ğŸ”’

test_munmap_back ğŸ”’

Exercise the MemoryManager logic for unmapping the back of a mapped region, validating that the still-mapped part of the region is still accessible. Regression test for #1188.

test_munmap_front ğŸ”’

Exercise the MemoryManager logic for unmapping the front of a mapped region, validating that the still-mapped part of the region is still accessible. Regression test for #1188.

test_munmap_middle ğŸ”’

Exercise the MemoryManager logic for unmapping the middle of a mapped region, validating that the still-mapped parts of the region is still accessible. Regression test for #1188.

validate_shadow_access ğŸ”’